refactor: move contents of './factories/__mocks__' folder to './factories/__mocks__/products'
