+++ app/back-end/src/factories/__mock__/products/create-product-controller-factory.ts
@@ -1,6 +1,6 @@
-import { CreateProductController } from '../../controllers/product/create-product-contoller'
-import { ProductRepositorySqlite } from '../../repositories/sqlite/__mock__/products-repository-sqlite'
-import { CreateProductService } from '../../services/product/create-product-service'
+import { CreateProductController } from '../../../controllers/product/create-product-contoller'
+import { ProductRepositorySqlite } from '../../../repositories/sqlite/__mock__/products-repository-sqlite'
+import { CreateProductService } from '../../../services/product/create-product-service'
 import { controllerAdapter } from 'vkrun'
 
 export const createProductControllerFactory = () => {
