+++ app/back-end/src/factories/__mock__/products/delete-product-controller-factory.ts
@@ -1,6 +1,6 @@
-import { DeleteProductController } from '../../controllers/product/delete-product-contoller'
-import { ProductRepositorySqlite } from '../../repositories/sqlite/__mock__/products-repository-sqlite'
-import { DeleteProductService } from '../../services/product/delete-product-service'
+import { DeleteProductController } from '../../../controllers/product/delete-product-contoller'
+import { ProductRepositorySqlite } from '../../../repositories/sqlite/__mock__/products-repository-sqlite'
+import { DeleteProductService } from '../../../services/product/delete-product-service'
 import { controllerAdapter } from 'vkrun'
 
 export const deleteProductControllerFactory = () => {
