+++ app/back-end/src/factories/__mock__/products/get-all-products-controller-factory.ts
@@ -1,7 +1,7 @@
-import { GetAllProductsService } from '../../services/product/get-all-products-service'
-import { ProductRepositorySqlite } from '../../repositories/sqlite/__mock__/products-repository-sqlite'
+import { GetAllProductsService } from '../../../services/product/get-all-products-service'
+import { ProductRepositorySqlite } from '../../../repositories/sqlite/__mock__/products-repository-sqlite'
 import { controllerAdapter } from 'vkrun'
-import { GetAllProductsController } from '../../controllers/product/get-all-products-contoller'
+import { GetAllProductsController } from '../../../controllers/product/get-all-products-contoller'
 
 export const getAllProductsControllerFactory = () => {
   const repository = new ProductRepositorySqlite()
