+++ app/back-end/src/factories/__mock__/products/get-product-controller-factory.ts
@@ -1,6 +1,6 @@
-import { GetProductController } from '../../controllers/product/get-product-contoller'
-import { ProductRepositorySqlite } from '../../repositories/sqlite/__mock__/products-repository-sqlite'
-import { GetProductService } from '../../services/product/get-product-service'
+import { GetProductController } from '../../../controllers/product/get-product-contoller'
+import { ProductRepositorySqlite } from '../../../repositories/sqlite/__mock__/products-repository-sqlite'
+import { GetProductService } from '../../../services/product/get-product-service'
 import { controllerAdapter } from 'vkrun'
 
 export const getProductControllerFactory = () => {
